feat!: adds support for ESM and Deno (#295)
