+++ .gitignore
@@ -1,8 +1,8 @@
 .idea
-build/
 .nyc_output
 node_modules
 .DS_Store
 package-lock.json
 ./test/fixtures/package.json
 coverage
+build
