--- .mocharc.json
@@ -1,6 +1,0 @@
-{
-  "spec": [
-    "build/test",
-    "test"
-  ]
-}
