+++ .nycrc
@@ -7,7 +7,7 @@
     "html",
     "text"
   ],
-  "lines": 100,
-  "branches": "97",
-  "statements": "100"
+  "lines": 99.5,
+  "branches": "98",
+  "statements": "99.5"
 }
