--- example.js
@@ -1,3 +1,0 @@
-var parser = require('./')
-var parse = parser('--foo "-bar"')
-console.log(parse)
