+++ package.json
@@ -2,15 +2,29 @@
   "name": "yargs-parser",
   "version": "18.1.3",
   "description": "the mighty option parser used by yargs",
-  "main": "build/index.js",
+  "main": "build/index.cjs",
+  "exports": {
+    "import": "./build/lib/index.js",
+    "require": "./build/index.cjs"
+  },
+  "type": "module",
+  "module": "./build/lib/index.js",
+  "types": "./build/index.cjs.d.ts",
   "scripts": {
-    "fix": "standardx --fix ./*.ts && standardx --fix **/*.ts",
-    "pretest": "npm run compile -- -p tsconfig.test.json",
-    "test": "c8 --reporter=text --reporter=html mocha test/*.js",
+    "check": "standardx '**/*.ts' && standardx '**/*.js' && standardx '**/*.cjs'",
+    "fix": "standardx --fix '**/*.ts' && standardx --fix '**/*.js' && standardx --fix '**/*.cjs'",
+    "pretest": "rimraf build && tsc -p tsconfig.test.json && cross-env NODE_ENV=test npm run build:cjs",
+    "test": "c8 --reporter=text --reporter=html mocha test/*.cjs",
+    "test:browser": "start-server-and-test 'serve ./ -p 8080' http://127.0.0.1:8080/package.json 'node ./test/browser/yargs-test.cjs'",
+    "pretest:typescript": "npm run pretest",
+    "test:typescript": "c8 mocha ./build/test/typescript/*.js",
     "posttest": "npm run check",
     "coverage": "c8 report --check-coverage",
-    "check": "standardx ./*.ts && standardx **/*.ts",
-    "compile": "rimraf build && tsc",
+    "precompile": "rimraf build",
+    "compile": "tsc",
+    "postcompile": "npm run build:cjs",
+    "build:cjs": "rollup -c",
+    "postbuild:cjs": "node scripts/replace-legacy-export.cjs",
     "prepare": "npm run compile"
   },
   "repository": {
@@ -36,33 +50,33 @@
     "@types/node": "^10.0.3",
     "@typescript-eslint/eslint-plugin": "^3.0.0",
     "@typescript-eslint/parser": "^3.0.0",
-    "c8": "^7.1.2",
+    "@wessberg/rollup-plugin-ts": "^1.2.28",
+    "c8": "^7.3.0",
     "chai": "^4.2.0",
+    "cross-env": "^7.0.2",
     "eslint": "^7.0.0",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-node": "^11.0.0",
     "gts": "^2.0.0-alpha.4",
     "mocha": "^8.0.0",
+    "puppeteer": "^5.2.1",
     "rimraf": "^3.0.2",
+    "rollup": "^2.22.1",
+    "serve": "^11.3.2",
     "standardx": "^5.0.0",
+    "start-server-and-test": "^1.11.2",
     "typescript": "^3.7.0"
   },
-  "dependencies": {
-    "camelcase": "^6.0.0",
-    "decamelize": "^4.0.0"
-  },
   "files": [
-    "index.js",
-    "build",
-    "lib/**/*.js"
+    "browser.js",
+    "build"
   ],
   "engines": {
     "node": ">=10"
   },
   "standardx": {
     "ignore": [
-      "build",
-      "example.js"
+      "build"
     ]
   }
 }
