+++ scripts/replace-legacy-export.cjs
@@ -1,0 +1,11 @@
+#!/usr/bin/env node
+'use strict'
+
+const { readFileSync, writeFileSync } = require('fs')
+
+// Cleanup the export statement in CJS typings file generated by rollup:
+const legacyTypings = require('path').resolve(__dirname, '../build/index.cjs.d.ts')
+const contents = readFileSync(legacyTypings, 'utf8').replace(
+  'export { yargsParser as default };', 'export = yargsParser;'
+)
+writeFileSync(legacyTypings, contents, 'utf8')
