+++ tsconfig.json
@@ -3,10 +3,12 @@
   "compilerOptions": {
     "outDir": "build",
     "rootDir": ".",
-    "sourceMap": false
-  },
+    "sourceMap": false,
+    "target": "es2017",
+    "moduleResolution": "node",
+    "module": "es2015"
+  },  
   "include": [
-    "./*.ts",
     "lib/**/*.ts"
   ]
 }
